# BackendV2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FUTCOL API",
        "description": "API para el sistema de gestión de torneos de fútbol FUTCOL. Esta API permite gestionar torneos, equipos, jugadores, amonestaciones y estadísticas.",
        "contact": {
            "name": "FUTCOL Support Team",
            "email": "admin@futcol.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "FUTCOL API Server"
        }
    ],
    "paths": {
...
        "/api/jugadores": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Listar todos los jugadores",
                "operationId": "dea173c276f606c29f946190cd3fbf0d",
                "responses": {
                    "200": {
                        "description": "Lista de jugadores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Crear un nuevo jugador",
                "operationId": "a6bca232b4957d71a1179acf4eba7097",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "n_documento",
                                    "fecha_nacimiento",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "n_documento": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    },
                                    "equipo_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jugador creado exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/jugadores/{id}": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Mostrar un jugador específico",
                "operationId": "38a1e42cd09832fbd502015f2a56d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del jugador"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Actualizar un jugador",
                "operationId": "43b9b9ac5454068b7a7befe507c8e46f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "n_documento": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "equipo_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jugador actualizado exitosamente"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Eliminar un jugador",
                "operationId": "ad3b63cc99eed2fea1490aca50426404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jugador eliminado exitosamente"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT en el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}